NXCM-1023: put in temp fix for dropping staged repos
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@4766 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-rest-api/src/main/java/org/sonatype/nexus/rest/repositories/RepositoryPlexusResource.java
@@ -23,9 +23,6 @@ import org.restlet.data.Status;
 import org.restlet.resource.ResourceException;
 import org.restlet.resource.Variant;
 import org.sonatype.nexus.configuration.ConfigurationException;
-import org.sonatype.nexus.configuration.model.CHttpProxySettings;
-import org.sonatype.nexus.configuration.model.CRemoteAuthentication;
-import org.sonatype.nexus.configuration.model.CRemoteConnectionSettings;
 import org.sonatype.nexus.configuration.model.RemoteSettingsUtil;
 import org.sonatype.nexus.proxy.NoSuchRepositoryException;
 import org.sonatype.nexus.proxy.StorageException;
@@ -289,8 +286,18 @@ public class RepositoryPlexusResource
         throws ResourceException
     {
         String repoId = this.getRepositoryId( request );
+        
         try
         {
+            //Can't mess with nexus-managed repos
+            //NOTE
+            //this should be in app logic, but dropping staging repositories
+            //should be allowed, so need to figure some way to handle both cases
+            if ( !getRepositoryRegistry().getRepository( repoId ).isUserManaged() )
+            {
+                throw new ResourceException( Status.CLIENT_ERROR_NOT_FOUND, "Repository Not Found" );
+            }
+            
             getNexus().deleteRepository( repoId );
 
             response.setStatus( Status.SUCCESS_NO_CONTENT );
